Renaming - change QuestionResponse to ContractQuestionResponse
