+++ services/app-web/src/pages/App/AppRoutes.tsx
@@ -32,7 +32,7 @@ import { useLocalStorage } from '../../hooks/useLocalStorage'
 import { recordJSException } from '../../otelHelpers'
 import { SubmissionSideNav } from '../SubmissionSideNav'
 import {
-    QuestionResponse,
+    ContractQuestionResponse,
     UploadContractResponse,
 } from '../QuestionResponse'
 import { GraphQLExplorer } from '../GraphQLExplorer/GraphQLExplorer'
@@ -49,7 +49,7 @@ import {
 import { EditStateAssign } from '../Settings/EditStateAssign/EditStateAssign'
 import { UploadContractQuestions, UploadRateQuestions } from '../QuestionResponse/UploadQuestions'
 import { RateSummarySideNav } from '../SubmissionSideNav/RateSummarySideNav'
-import { RateQuestionResponse } from '../QuestionResponse/RateQuestionResponse'
+import { RateQuestionResponse } from '../QuestionResponse/QuestionResponseSummary/RateQuestionResponse'
 import { UploadRateResponse } from '../QuestionResponse/UploadResponse/UploadRateResponse'
 
 function componentForAuthMode(
@@ -138,7 +138,7 @@ const StateUserRoutes = ({
                         path={
                             RoutesRecord.SUBMISSIONS_CONTRACT_QUESTIONS_AND_ANSWERS
                         }
-                        element={<QuestionResponse />}
+                        element={<ContractQuestionResponse />}
                     />
                     <Route
                         element={<UploadContractResponse />}
@@ -231,7 +231,7 @@ const CMSUserRoutes = ({
                         path={
                             RoutesRecord.SUBMISSIONS_CONTRACT_QUESTIONS_AND_ANSWERS
                         }
-                        element={<QuestionResponse />}
+                        element={<ContractQuestionResponse />}
                     />
                     <Route
                         path={RoutesRecord.SUBMISSIONS_UPLOAD_CONTRACT_QUESTION}
