+++ services/app-web/src/pages/QuestionResponse/UploadQuestions/UploadRateQuestions.test.tsx
@@ -23,7 +23,7 @@ import {
 } from '../../../testHelpers/apolloMocks/questionResponseGQLMock'
 import { Location } from 'react-router-dom'
 import { UploadRateQuestions } from './UploadRateQuestions'
-import { RateQuestionResponse } from '../RateQuestionResponse'
+import { RateQuestionResponse } from '../QuestionResponseSummary/RateQuestionResponse'
 
 describe('UploadRateQuestions', () => {
     afterEach(() => {
