+++ services/app-web/src/pages/QuestionResponse/UploadResponse/index.ts
@@ -1,2 +1,3 @@
 export { UploadContractResponse } from './UploadContractResponse'
+export { UploadRateResponse } from './UploadRateResponse'
 export { UploadResponseForm } from './UploadResponseForm'
